Adding API call to fetch active teams.
