+++ shared-test/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/test/TestDataModule.kt
@@ -4,7 +4,7 @@ import com.adammcneilly.pocketleague.data.event.EventService
 import com.adammcneilly.pocketleague.data.event.test.FakeEventService
 import com.adammcneilly.pocketleague.data.game.GameService
 import com.adammcneilly.pocketleague.data.match.MatchService
-import com.adammcneilly.pocketleague.data.team.DemoTeamService
+import com.adammcneilly.pocketleague.data.team.OctaneGGTeamService
 import com.adammcneilly.pocketleague.data.team.TeamService
 import com.adammcneilly.pocketleague.shared.di.DataModule
 
@@ -20,5 +20,5 @@ class TestDataModule : DataModule {
 
     override val gameService: GameService = GameService.provideDefault()
 
-    override val teamService: TeamService = DemoTeamService()
+    override val teamService: TeamService = OctaneGGTeamService()
 }
