+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/di/ProdDataModule.kt
@@ -3,7 +3,7 @@ package com.adammcneilly.pocketleague.shared.di
 import com.adammcneilly.pocketleague.data.event.EventService
 import com.adammcneilly.pocketleague.data.game.GameService
 import com.adammcneilly.pocketleague.data.match.MatchService
-import com.adammcneilly.pocketleague.data.team.DemoTeamService
+import com.adammcneilly.pocketleague.data.team.OctaneGGTeamService
 import com.adammcneilly.pocketleague.data.team.TeamService
 
 /**
@@ -24,6 +24,6 @@ class ProdDataModule : DataModule {
         GameService.provideDefault()
     }
     override val teamService: TeamService by lazy {
-        DemoTeamService()
+        OctaneGGTeamService()
     }
 }
